% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igfetchr.R
\name{ig_close_session}
\alias{ig_close_session}
\title{Close session / logout (alias)}
\usage{
ig_close_session(auth, api_key = NULL)
}
\arguments{
\item{auth}{Authentication list from ig_auth()}

\item{api_key}{API key (optional)}
}
\value{
Logical TRUE if the session was closed successfully.
}
\description{
Convenience wrapper for ig_logout()
}
