% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igfetchr.R
\name{ig_logout}
\alias{ig_logout}
\title{Close session / logout}
\usage{
ig_logout(auth, api_key = NULL)
}
\arguments{
\item{auth}{Authentication list from ig_auth()}

\item{api_key}{API key (optional)}
}
\value{
Logical TRUE if the session was closed successfully.
}
\description{
Close the authenticated session.
}
\examples{
\dontrun{
auth <- ig_auth("user", "pass", "api_key")
ig_logout(auth)
}
}
