% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igfetchr.R
\docType{package}
\name{igfetchr-package}
\alias{igfetchr-package}
\title{igfetchr: Lightweight IG Trading REST API wrapper}
\description{
Helper functions to authenticate and fetch market/account data from the IG Trading REST API.
All exported functions return tibbles where appropriate. This package is for data access only
and not financial advice. Trading CFDs and spread bets carries a high risk of losing money.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sawsimeon/igfetchr}
  \item Report bugs at \url{https://github.com/sawsimeon/igfetchr/issues}
}

}
\author{
\strong{Maintainer}: Saw Simeon \email{saw.simeon.brown@gmail.com}

}
\keyword{internal}
