% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igfetchr.R
\docType{package}
\name{igfetchr}
\alias{igfetchr}
\title{igfetchr: Lightweight IG Trading REST API wrapper}
\description{
Helper functions to authenticate and fetch market/account data from the IG Trading
REST API. All exported functions return tibbles where appropriate. This package is
not financial advice. Trading CFDs and spread bets carries high risk of losing money.
Key functions: `ig_auth` to authenticate, `ig_execute_trade` to place trades.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sawsimeon/igfetchr}
  \item \url{https://sawsimeon.github.io/igfetchr/}
  \item Report bugs at \url{https://github.com/sawsimeon/igfetchr/issues}
}

}
\author{
\strong{Maintainer}: Saw Simeon \email{saw.s@ku.th}

}
